requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
numpy>=1.24.3
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
glfw>=2.5.9
Pillow>=10.0.0
